pyproject.toml
[project]
name = "literesearch"
version = "0.1.0"
description = "AI-powered research tool built with Python"
authors = [
    {name = "Richard Wang"},
]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "duckduckgo-search>=8.0.4",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-openai>=0.3.26",
    "langfuse>=3.0.6",
    "lxml>=6.0.0",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "tavily-python>=0.7.8",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
]
